refactor(bridge): decompose createBridge() into per-feature composition modules
Split the ~561-line bridge.ts createBridge() closure into a thin
composition root plus per-feature wiring modules under composition/,
following the precedent in apps/vscode-plugin/src/composition/:
a BridgeSharedDeps bag, per-feature register() modules, handles flowing
forward and session hooks flowing backward.

The cross-feature session lifecycle (session/register, session/dispose)
is resolved with a SessionHooks contract — the bridge counterpart of the
VS Code host's EditorSessionParticipant — so each feature keeps its own
per-session state (template watchers, script tabs, code-link maps)
without the editor-session feature importing it.

Pure code move: every method is registered exactly once and all specs
pass unmodified.

Closes #1102
